# Conflicts:
#	actors/src/main/scala/org/loopring/lightcone/actors/ethereum/event/TokenBurnRateEventExtractor.scala
#	proto/src/main/protobuf/actors_messages.proto

+++ persistence/src/main/scala/org/loopring/lightcone/persistence/dals/MissingBlocksRecordDal.scala
@@ -23,8 +23,10 @@ import scala.concurrent._
 
 trait MissingBlocksRecordDal
     extends BaseDalImpl[MissingBlocksRecordTable, MissingBlocksRecord] {
-  def saveMissingBlock(record: MissingBlocksRecord): Future[Int]
+  // will return saved sequenceId
+  def saveMissingBlock(record: MissingBlocksRecord): Future[Long]
 
+  // order by insert sequence, get the earliest one
   def getOldestOne(): Future[Option[MissingBlocksRecord]]
 
   def updateProgress(
